+++ src/main/java/com/jluzh/sysmanagement/domain/repository/AreaRepository.java
@@ -16,13 +16,13 @@ public interface AreaRepository {
 
 	List<Area> selectList(Area area);
 
-	int deleteByPrimaryKey(Long id);
+	int deleteByPrimaryKey(Integer id);
 
 	int insert(Area record);
 
 	int insertSelective(Area record);
 
-	Area selectByPrimaryKey(Long id);
+	Area selectByPrimaryKey(Integer id);
 
 	int updateByPrimaryKeySelective(Area record);
