+++ src/main/java/com/jluzh/sysmanagement/infra/mapper/AreaMapper.java
@@ -14,13 +14,13 @@ import java.util.List;
 public interface AreaMapper {
     List<Area> selectList(Area area);
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Area record);
 
     int insertSelective(Area record);
 
-    Area selectByPrimaryKey(Long id);
+    Area selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(Area record);
